separate v3 dispatcher by icx / ise / rep
