+++ iconrpcserver/dispatcher/v3/__init__.py
@@ -13,4 +13,6 @@
 # limitations under the License.
 
 from .version3 import *
-
+from .icx import *
+from .ise import *
+from .rep import *
